change logic following enum changes
