+++ indexers/gemini-3h/staking/src/mappings/mappingHandlers.ts
@@ -1,30 +1,15 @@
 import { SubstrateEvent, SubstrateExtrinsic } from "@subql/types";
 import assert from "assert";
+import { Deposit, Nominator, Withdrawal } from "../types";
+import * as db from "./db";
 import {
-  Account,
-  Bundle,
-  BundleAuthor,
-  Deposit,
   DepositStatus,
-  Domain,
-  DomainBlock,
-  DomainEpoch,
-  DomainRuntime,
-  Nominator,
   NominatorPendingAction,
   NominatorStatus,
-  Operator,
   OperatorPendingAction,
   OperatorStatus,
-  Reward,
-  Stats,
-  StatsPerAccount,
-  StatsPerDomain,
-  StatsPerNominator,
-  StatsPerOperator,
-  Withdrawal,
   WithdrawalStatus,
-} from "../types";
+} from "./models";
 
 export const SHARES_CALCULATION_MULTIPLIER = BigInt(1000000000000);
 
@@ -36,13 +21,18 @@ export async function handleRegisterOperatorCall(
       extrinsic.extrinsic.method
     } extrinsic found at block ${extrinsic.block.block.header.number.toString()}`
   );
-  const signer = extrinsic.extrinsic.signer.toString();
-  const blockNumber = extrinsic.block.block.header.number.toNumber();
   const {
-    events,
+    block: {
+      timestamp,
+      block: {
+        header: { number },
+      },
+    },
     extrinsic: {
       method: { args },
+      signer,
     },
+    events,
   } = extrinsic;
   const domainId = String(args[0]);
   const amount = BigInt(args[1].toString());
@@ -68,21 +58,24 @@ export async function handleRegisterOperatorCall(
 
   assert(domainId === domainId_storageFee, "domainId mismatch");
   assert(operatorId === operatorId_storageFee, "operatorId mismatch");
-  assert(signer === nominatorId_storageFee, "signer mismatch");
+  assert(signer.toString() === nominatorId_storageFee, "signer mismatch");
   assert(
     amount_storageFee === (BigInt(amount) * BigInt(20)) / BigInt(100),
     `Storage fee (${amount_storageFee}) should be 20% of the amount (${amount})`
   );
 
-  const domain = await checkAndGetDomain(domainId, blockNumber);
+  const domain = await db.checkAndGetDomain(domainId, number.toNumber());
   domain.totalDeposits += amount;
-  domain.updatedAt = blockNumber;
+  domain.updatedAt = number.toNumber();
 
-  const account = await checkAndGetAccount(signer, blockNumber);
+  const account = await db.checkAndGetAccount(
+    signer.toString(),
+    number.toNumber()
+  );
   account.totalDeposits += amount;
-  account.updatedAt = blockNumber;
+  account.updatedAt = number.toNumber();
 
-  const operator = await checkAndGetOperator(operatorId, blockNumber);
+  const operator = await db.checkAndGetOperator(operatorId, number.toNumber());
   operator.domainId = domain.id;
   operator.accountId = account.id;
   operator.signingKey = String(signingKey);
@@ -90,25 +83,25 @@ export async function handleRegisterOperatorCall(
   operator.nominationTax = Number(nominationTax);
   operator.totalDeposits += amount;
   operator.pendingAction = OperatorPendingAction.PENDING_REGISTRATION;
-  operator.updatedAt = blockNumber;
+  operator.updatedAt = number.toNumber();
 
-  const nominator = await checkAndGetNominator(
-    signer,
+  const nominator = await db.checkAndGetNominator(
+    signer.toString(),
     domain.id,
     operator.id,
-    blockNumber
+    number.toNumber()
   );
   nominator.totalDeposits += amount;
   nominator.pendingAction = NominatorPendingAction.PENDING_EPOCH_CHANGE;
-  nominator.updatedAt = blockNumber;
+  nominator.updatedAt = number.toNumber();
 
-  const deposit = await checkAndGetDeposit(
-    signer,
+  const deposit = await db.checkAndGetDeposit(
+    signer.toString(),
     domain.id,
     operator.id,
     amount - amount_storageFee,
     amount_storageFee,
-    blockNumber
+    number.toNumber()
   );
   deposit.epochDepositedAt = domain.completedEpoch ?? 0;
   deposit.domainBlockNumberDepositedAt = domain.lastDomainBlockNumber ?? 0;
@@ -148,18 +141,18 @@ export async function handleNominateOperatorCall(
   );
   const totalAmount = amount + storageFeeDeposit;
 
-  const account = await checkAndGetAccount(signer, blockNumber);
-  const operator = await checkAndGetOperator(operatorId, blockNumber);
-  const domain = await checkAndGetDomain(operator.domainId, blockNumber);
+  const account = await db.checkAndGetAccount(signer, blockNumber);
+  const operator = await db.checkAndGetOperator(operatorId, blockNumber);
+  const domain = await db.checkAndGetDomain(operator.domainId, blockNumber);
 
-  const nominator = await checkAndGetNominator(
+  const nominator = await db.checkAndGetNominator(
     signer,
     domain.id,
     operator.id,
     blockNumber
   );
 
-  const deposit = await checkAndGetDeposit(
+  const deposit = await db.checkAndGetDeposit(
     signer,
     domain.id,
     operator.id,
@@ -220,14 +213,14 @@ export async function handleDeregisterOperatorCall(
     return;
   }
 
-  const operator = await checkAndGetOperator(operatorId, blockNumber);
+  const operator = await db.checkAndGetOperator(operatorId, blockNumber);
   operator.currentTotalStake = BigInt(0);
   operator.currentStorageFeeDeposit = BigInt(0);
   operator.status = OperatorStatus.DEREGISTERED;
   operator.updatedAt = blockNumber;
 
-  const domain = await checkAndGetDomain(operator.domainId, blockNumber);
-  const account = await checkAndGetAccount(operator.accountId, blockNumber);
+  const domain = await db.checkAndGetDomain(operator.domainId, blockNumber);
+  const account = await db.checkAndGetAccount(operator.accountId, blockNumber);
 
   const nominators = await Nominator.getByOperatorId(operator.id);
   const activeNominators =
@@ -242,7 +235,7 @@ export async function handleDeregisterOperatorCall(
       const estimatedAmount =
         (operator.currentSharePrice * n.knownShares) /
         SHARES_CALCULATION_MULTIPLIER;
-      const withdrawal = await checkAndGetWithdrawal(
+      const withdrawal = await db.checkAndGetWithdrawal(
         account.id,
         domain.id,
         operator.id,
@@ -273,7 +266,7 @@ export async function handleOperatorSlashedEvent(
   const operatorId = event.event.data[0].toString();
 
   // Query operator details
-  const operator = await checkAndGetOperator(operatorId, blockNumber);
+  const operator = await db.checkAndGetOperator(operatorId, blockNumber);
   operator.currentTotalStake = BigInt(0);
   operator.currentStorageFeeDeposit = BigInt(0);
   operator.status = OperatorStatus.SLASHED;
@@ -322,12 +315,12 @@ export async function handleOperatorTaxCollectedEvent(
   const blockNumber = block.block.header.number.toNumber();
   const taxAmount = BigInt(tax.toString());
 
-  const operator = await checkAndGetOperator(
+  const operator = await db.checkAndGetOperator(
     operatorId.toString(),
     blockNumber
   );
-  const account = await checkAndGetAccount(operator.accountId, blockNumber);
-  const domain = await checkAndGetDomain(operator.domainId, blockNumber);
+  const account = await db.checkAndGetAccount(operator.accountId, blockNumber);
+  const domain = await db.checkAndGetDomain(operator.domainId, blockNumber);
 
   operator.totalTaxCollected += taxAmount;
   operator.updatedAt = blockNumber;
@@ -352,11 +345,11 @@ export async function handleOperatorRewardedEvent(
   } = event;
   const amount = BigInt(reward.toString());
 
-  const operator = await checkAndGetOperator(
+  const operator = await db.checkAndGetOperator(
     operatorId.toString(),
     block.block.header.number.toNumber()
   );
-  const domain = await checkAndGetDomain(
+  const domain = await db.checkAndGetDomain(
     operator.domainId,
     block.block.header.number.toNumber()
   );
@@ -364,7 +357,7 @@ export async function handleOperatorRewardedEvent(
   operator.totalRewardsCollected += amount;
   domain.totalRewardsCollected += amount;
 
-  const rewardEvent = await checkAndGetReward(
+  const rewardEvent = await db.checkAndGetReward(
     `${operatorId}-${block.block.header.number.toNumber()}`,
     block.block.header.number.toNumber()
   );
@@ -390,13 +383,13 @@ export async function handleWithdrawStakeCall(
   const blockNumber = extrinsic.block.block.header.number.toNumber();
   const sharesBigInt = BigInt(shares.toString());
 
-  const account = await checkAndGetAccount(address, blockNumber);
-  const operator = await checkAndGetOperator(
+  const account = await db.checkAndGetAccount(address, blockNumber);
+  const operator = await db.checkAndGetOperator(
     operatorId.toString(),
     blockNumber
   );
-  const domain = await checkAndGetDomain(operator.domainId, blockNumber);
-  const nominator = await checkAndGetNominator(
+  const domain = await db.checkAndGetDomain(operator.domainId, blockNumber);
+  const nominator = await db.checkAndGetNominator(
     address,
     domain.id,
     operator.id,
@@ -405,7 +398,7 @@ export async function handleWithdrawStakeCall(
 
   const estimatedAmount =
     (operator.currentSharePrice * sharesBigInt) / SHARES_CALCULATION_MULTIPLIER;
-  const withdrawal = await checkAndGetWithdrawal(
+  const withdrawal = await db.checkAndGetWithdrawal(
     account.id,
     domain.id,
     operator.id,
@@ -457,7 +450,7 @@ export async function handleUnlockFundsCall(
   const [operatorId, shares] = extrinsic.extrinsic.method.args;
 
   // Query operator details
-  const operator = await checkAndGetOperator(
+  const operator = await db.checkAndGetOperator(
     operatorId.toString(),
     blockNumber
   );
@@ -468,7 +461,7 @@ export async function handleUnlockFundsCall(
     (operator.currentSharePrice * sharesBigInt) / SHARES_CALCULATION_MULTIPLIER;
 
   // Fetch or create Withdrawal entity
-  const withdrawal = await checkAndGetWithdrawal(
+  const withdrawal = await db.checkAndGetWithdrawal(
     signer,
     operator.domainId,
     operator.id,
@@ -483,9 +476,9 @@ export async function handleUnlockFundsCall(
   withdrawal.updatedAt = blockNumber;
 
   // Update related entities
-  const account = await checkAndGetAccount(signer, blockNumber);
-  const domain = await checkAndGetDomain(operator.domainId, blockNumber);
-  const nominator = await checkAndGetNominator(
+  const account = await db.checkAndGetAccount(signer, blockNumber);
+  const domain = await db.checkAndGetDomain(operator.domainId, blockNumber);
+  const nominator = await db.checkAndGetNominator(
     signer,
     domain.id,
     operator.id,
@@ -545,7 +538,7 @@ export async function handleDomainInstantiatedEvent(
   const completedEpoch = Number(_completedEpochIndex?.toString() ?? 0);
   const blockNumber = event.block.block.header.number.toNumber();
 
-  const domain = await checkAndGetDomain(domainId, blockNumber);
+  const domain = await db.checkAndGetDomain(domainId, blockNumber);
   domain.completedEpoch = completedEpoch;
   domain.updatedAt = blockNumber;
 
@@ -579,7 +572,7 @@ export async function handleDomainEpochCompletedEvent(
   };
 
   // Fetch or create Domain entity
-  const domain = await checkAndGetDomain(domainId.toString(), blockNumber);
+  const domain = await db.checkAndGetDomain(domainId.toString(), blockNumber);
   domain.completedEpoch = Number(epochIndex);
   domain.updatedAt = blockNumber;
 
@@ -841,12 +834,12 @@ export async function handleOperatorUnlockedEvent(
     block,
   } = event;
   const blockNumber = block.block.header.number.toNumber();
-  const operator = await checkAndGetOperator(
+  const operator = await db.checkAndGetOperator(
     operatorId.toString(),
     blockNumber
   );
-  const domain = await checkAndGetDomain(operator.domainId, blockNumber);
-  const account = await checkAndGetAccount(operator.accountId, blockNumber);
+  const domain = await db.checkAndGetDomain(operator.domainId, blockNumber);
+  const account = await db.checkAndGetAccount(operator.accountId, blockNumber);
 
   operator.pendingAction = OperatorPendingAction.NO_ACTION_REQUIRED;
   operator.updatedAt = blockNumber;
@@ -893,13 +886,16 @@ export async function handleFundsUnlockedEvent(
   const blockNumber = block.block.header.number.toNumber();
   const amountBigInt = BigInt(amount.toString());
 
-  const operator = await checkAndGetOperator(
+  const operator = await db.checkAndGetOperator(
     operatorId.toString(),
     blockNumber
   );
-  const domain = await checkAndGetDomain(operator.domainId, blockNumber);
-  const account = await checkAndGetAccount(nominatorId.toString(), blockNumber);
-  const nominator = await checkAndGetNominator(
+  const domain = await db.checkAndGetDomain(operator.domainId, blockNumber);
+  const account = await db.checkAndGetAccount(
+    nominatorId.toString(),
+    blockNumber
+  );
+  const nominator = await db.checkAndGetNominator(
     nominatorId.toString(),
     domain.id,
     operator.id,
@@ -982,542 +978,4 @@ export async function handleStorageFeeUnlockedEvent(
   event: SubstrateEvent
 ): Promise<void> {
   // Implementation for handleStorageFeeUnlockedEvent
-}
-
-// Helper for Domain
-export async function checkAndGetDomain(
-  domainId: string,
-  blockNumber: number
-): Promise<Domain> {
-  let domain = await Domain.get(domainId);
-  if (!domain) {
-    domain = Domain.create({
-      id: domainId.toLowerCase(),
-      sortId: Number(0),
-      accountId: "",
-      name: "",
-      runtimeId: 0,
-      runtime: DomainRuntime.EVM,
-      runtimeInfo: "",
-      completedEpoch: 0,
-      lastDomainBlockNumber: 0,
-      totalDeposits: BigInt(0),
-      totalEstimatedWithdrawals: BigInt(0),
-      totalWithdrawals: BigInt(0),
-      totalTaxCollected: BigInt(0),
-      totalRewardsCollected: BigInt(0),
-      totalTransfersIn: BigInt(0),
-      transfersInCount: 0,
-      totalTransfersOut: BigInt(0),
-      transfersOutCount: 0,
-      totalRejectedTransfersClaimed: BigInt(0),
-      rejectedTransfersClaimedCount: 0,
-      totalTransfersRejected: BigInt(0),
-      transfersRejectedCount: 0,
-      totalVolume: BigInt(0),
-      totalConsensusStorageFee: BigInt(0),
-      totalDomainExecutionFee: BigInt(0),
-      totalBurnedBalance: BigInt(0),
-      currentTotalStake: BigInt(0),
-      currentStorageFeeDeposit: BigInt(0),
-      currentTotalShares: BigInt(0),
-      currentSharePrice: BigInt(0),
-      accumulatedEpochStake: BigInt(0),
-      accumulatedEpochStorageFeeDeposit: BigInt(0),
-      accumulatedEpochRewards: BigInt(0),
-      accumulatedEpochShares: BigInt(0),
-      bundleCount: 0,
-      currentEpochDuration: BigInt(0),
-      lastEpochDuration: BigInt(0),
-      last6EpochsDuration: BigInt(0),
-      last144EpochDuration: BigInt(0),
-      last1kEpochDuration: BigInt(0),
-      lastBundleAt: 0,
-      createdAt: blockNumber,
-      updatedAt: blockNumber,
-    });
-  }
-  return domain;
-}
-
-// Helper for Account
-export async function checkAndGetAccount(
-  id: string,
-  blockNumber: number
-): Promise<Account> {
-  let account = await Account.get(id.toLowerCase());
-  if (!account) {
-    account = Account.create({
-      id: id.toLowerCase(),
-      totalDeposits: BigInt(0),
-      totalEstimatedWithdrawals: BigInt(0),
-      totalWithdrawals: BigInt(0),
-      totalTaxCollected: BigInt(0),
-      currentTotalStake: BigInt(0),
-      currentStorageFeeDeposit: BigInt(0),
-      currentTotalShares: BigInt(0),
-      currentSharePrice: BigInt(0),
-      accumulatedEpochStake: BigInt(0),
-      accumulatedEpochStorageFeeDeposit: BigInt(0),
-      accumulatedEpochShares: BigInt(0),
-      createdAt: blockNumber,
-      updatedAt: blockNumber,
-    });
-  }
-  return account;
-}
-
-// Helper for Operator
-export async function checkAndGetOperator(
-  operatorId: string,
-  blockNumber: number
-): Promise<Operator> {
-  const id = operatorId.toLowerCase();
-  let operator = await Operator.get(id);
-  if (!operator) {
-    operator = Operator.create({
-      id,
-      sortId: Number(operatorId),
-      accountId: "",
-      domainId: "",
-      signingKey: "",
-      minimumNominatorStake: BigInt(0),
-      nominationTax: 0,
-      currentTotalStake: BigInt(0),
-      currentStorageFeeDeposit: BigInt(0),
-      currentEpochRewards: BigInt(0),
-      currentTotalShares: BigInt(0),
-      currentSharePrice: BigInt(0),
-      rawStatus: "",
-      totalDeposits: BigInt(0),
-      totalEstimatedWithdrawals: BigInt(0),
-      totalWithdrawals: BigInt(0),
-      totalTaxCollected: BigInt(0),
-      totalRewardsCollected: BigInt(0),
-      totalTransfersIn: BigInt(0),
-      transfersInCount: 0,
-      totalTransfersOut: BigInt(0),
-      transfersOutCount: 0,
-      totalRejectedTransfersClaimed: BigInt(0),
-      rejectedTransfersClaimedCount: 0,
-      totalTransfersRejected: BigInt(0),
-      transfersRejectedCount: 0,
-      totalVolume: BigInt(0),
-      totalConsensusStorageFee: BigInt(0),
-      totalDomainExecutionFee: BigInt(0),
-      totalBurnedBalance: BigInt(0),
-      accumulatedEpochStake: BigInt(0),
-      accumulatedEpochStorageFeeDeposit: BigInt(0),
-      accumulatedEpochRewards: BigInt(0),
-      accumulatedEpochShares: BigInt(0),
-      activeEpochCount: 0,
-      bundleCount: 0,
-      status: OperatorStatus.PENDING,
-      pendingAction: OperatorPendingAction.NO_ACTION_REQUIRED,
-      lastBundleAt: 0,
-      createdAt: blockNumber,
-      updatedAt: blockNumber,
-    });
-  }
-  return operator;
-}
-
-// Helper for DomainBlock
-export async function checkAndGetDomainBlock(
-  id: string,
-  blockNumber: number
-): Promise<DomainBlock> {
-  let domainBlock = await DomainBlock.get(id.toLowerCase());
-  if (!domainBlock) {
-    domainBlock = DomainBlock.create({
-      id: id.toLowerCase(),
-      domainId: "",
-      domainEpochId: "",
-      blockNumber: 0,
-      blockHash: "",
-      extrinsicRoot: "",
-      epoch: 0,
-      consensusBlockNumber: 0,
-      consensusBlockHash: "",
-      timestamp: new Date(),
-      createdAt: blockNumber,
-      updatedAt: blockNumber,
-    });
-  }
-  return domainBlock;
-}
-
-// Helper for DomainEpoch
-export async function checkAndGetDomainEpoch(
-  id: string,
-  blockNumber: number
-): Promise<DomainEpoch> {
-  let domainEpoch = await DomainEpoch.get(id.toLowerCase());
-  if (!domainEpoch) {
-    domainEpoch = DomainEpoch.create({
-      id: id.toLowerCase(),
-      epoch: 0,
-      domainId: "",
-      blockNumberStart: 0,
-      blockNumberEnd: 0,
-      blockCount: 0,
-      timestampStart: new Date(),
-      timestampEnd: new Date(),
-      epochDuration: BigInt(0),
-      consensusBlockNumberStart: 0,
-      consensusBlockNumberEnd: 0,
-      consensusBlockHashStart: "",
-      consensusBlockHashEnd: "",
-      createdAt: blockNumber,
-      updatedAt: blockNumber,
-    });
-  }
-  return domainEpoch;
-}
-
-// Helper for Bundle
-export async function checkAndGetBundle(
-  id: string,
-  blockNumber: number
-): Promise<Bundle> {
-  let bundle = await Bundle.get(id.toLowerCase());
-  if (!bundle) {
-    bundle = Bundle.create({
-      id: id.toLowerCase(),
-      domainId: "",
-      domainBlockId: "",
-      domainEpochId: "",
-      domainBlockNumber: 0,
-      domainBlockHash: "",
-      domainBlockExtrinsicRoot: "",
-      epoch: 0,
-      consensusBlockNumber: 0,
-      consensusBlockHash: "",
-      totalTransfersIn: BigInt(0),
-      transfersInCount: 0,
-      totalTransfersOut: BigInt(0),
-      transfersOutCount: 0,
-      totalRejectedTransfersClaimed: BigInt(0),
-      rejectedTransfersClaimedCount: 0,
-      totalTransfersRejected: BigInt(0),
-      transfersRejectedCount: 0,
-      totalVolume: BigInt(0),
-      consensusStorageFee: BigInt(0),
-      domainExecutionFee: BigInt(0),
-      burnedBalance: BigInt(0),
-    });
-  }
-  return bundle;
-}
-
-// Helper for BundleAuthor
-export async function checkAndGetBundleAuthor(
-  id: string,
-  blockNumber: number
-): Promise<BundleAuthor> {
-  let bundleAuthor = await BundleAuthor.get(id.toLowerCase());
-  if (!bundleAuthor) {
-    bundleAuthor = BundleAuthor.create({
-      id: id.toLowerCase(),
-      domainId: "",
-      accountId: "",
-      operatorId: "",
-      bundleId: "",
-      domainBlockId: "",
-      domainEpochId: "",
-      epoch: 0,
-    });
-  }
-  return bundleAuthor;
-}
-
-// Helper for Nominator
-export async function checkAndGetNominator(
-  accountId: string,
-  domainId: string,
-  operatorId: string,
-  blockNumber: number
-): Promise<Nominator> {
-  const id = `${operatorId}-${accountId}`.toLowerCase();
-  let nominator = await Nominator.get(id);
-  if (!nominator) {
-    nominator = Nominator.create({
-      id,
-      accountId,
-      domainId,
-      operatorId,
-      knownShares: BigInt(0),
-      knownStorageFeeDeposit: BigInt(0),
-      pendingAmount: BigInt(0),
-      pendingStorageFeeDeposit: BigInt(0),
-      pendingEffectiveDomainEpoch: 0,
-      totalWithdrawalAmounts: BigInt(0),
-      totalStorageFeeRefund: BigInt(0),
-      unlockAtConfirmedDomainBlockNumber: [],
-      pendingShares: BigInt(0),
-      pendingStorageFeeRefund: BigInt(0),
-      totalDeposits: BigInt(0),
-      totalEstimatedWithdrawals: BigInt(0),
-      totalWithdrawals: BigInt(0),
-      totalDepositsCount: 0,
-      totalWithdrawalsCount: 0,
-      currentTotalStake: BigInt(0),
-      currentStorageFeeDeposit: BigInt(0),
-      currentTotalShares: BigInt(0),
-      currentSharePrice: BigInt(0),
-      accumulatedEpochStake: BigInt(0),
-      accumulatedEpochStorageFeeDeposit: BigInt(0),
-      accumulatedEpochShares: BigInt(0),
-      activeEpochCount: 0,
-      status: NominatorStatus.PENDING,
-      pendingAction: NominatorPendingAction.NO_ACTION_REQUIRED,
-      createdAt: blockNumber,
-      updatedAt: blockNumber,
-    });
-  }
-  return nominator;
-}
-
-// Helper for Deposit
-export async function checkAndGetDeposit(
-  accountId: string,
-  domainId: string,
-  operatorId: string,
-  amount: bigint,
-  storageFeeDeposit: bigint,
-  blockNumber: number
-): Promise<Deposit> {
-  const id = `${operatorId}-${accountId}-${blockNumber}`.toLowerCase();
-  const nominatorId = `${operatorId}-${accountId}`.toLowerCase();
-  const totalAmount = amount + storageFeeDeposit;
-  let deposit = await Deposit.get(id);
-  if (!deposit) {
-    deposit = Deposit.create({
-      id,
-      accountId,
-      domainId,
-      operatorId,
-      nominatorId,
-      amount,
-      storageFeeDeposit,
-      totalAmount,
-      totalWithdrawn: BigInt(0),
-      status: DepositStatus.PENDING,
-      timestamp: new Date(),
-      extrinsicHash: "",
-      epochDepositedAt: 0,
-      domainBlockNumberDepositedAt: 0,
-      createdAt: blockNumber,
-      stakedAt: 0,
-      updatedAt: blockNumber,
-    });
-  }
-  return deposit;
-}
-
-// Helper for Withdrawal
-export async function checkAndGetWithdrawal(
-  accountId: string,
-  domainId: string,
-  operatorId: string,
-  nominatorId: string,
-  blockNumber: number
-): Promise<Withdrawal> {
-  const id = `${operatorId}-${accountId}-${blockNumber}`.toLowerCase();
-  let withdrawal = await Withdrawal.get(id);
-  if (!withdrawal) {
-    withdrawal = Withdrawal.create({
-      id,
-      accountId,
-      domainId,
-      operatorId,
-      nominatorId,
-      shares: BigInt(0),
-      estimatedAmount: BigInt(0),
-      unlockedAmount: BigInt(0),
-      unlockedStorageFee: BigInt(0),
-      totalAmount: BigInt(0),
-      status: WithdrawalStatus.PENDING_LOCK,
-      timestamp: new Date(),
-      withdrawExtrinsicHash: "",
-      unlockExtrinsicHash: "",
-      epochWithdrawalRequestedAt: 0,
-      domainBlockNumberWithdrawalRequestedAt: 0,
-      createdAt: blockNumber,
-      readyAt: 0,
-      unlockedAt: 0,
-      updatedAt: blockNumber,
-    });
-  }
-  return withdrawal;
-}
-
-// Helper for Reward
-export async function checkAndGetReward(
-  id: string,
-  blockNumber: number
-): Promise<Reward> {
-  let reward = await Reward.get(id.toLowerCase());
-  if (!reward) {
-    reward = Reward.create({
-      id: id.toLowerCase(),
-      domainId: "",
-      operatorId: "",
-      amount: BigInt(0),
-      timestamp: new Date(),
-      blockNumber: 0,
-      extrinsicHash: "",
-    });
-  }
-  return reward;
-}
-
-// Helper for Stats
-export async function checkAndGetStats(
-  id: string,
-  blockNumber: number
-): Promise<Stats> {
-  let stats = await Stats.get(id.toLowerCase());
-  if (!stats) {
-    stats = Stats.create({
-      id: id.toLowerCase(),
-      blockNumber: 0,
-      totalStaked: BigInt(0),
-      totalTaxCollected: BigInt(0),
-      totalRewardsCollected: BigInt(0),
-      totalDeposits: BigInt(0),
-      totalWithdrawals: BigInt(0),
-      totalShares: BigInt(0),
-      currentSharePrice: BigInt(0),
-      allTimeHighStaked: BigInt(0),
-      allTimeHighSharePrice: BigInt(0),
-      domainsCount: 0,
-      operatorsCount: 0,
-      activeOperatorsCount: 0,
-      slashedOperatorsCount: 0,
-      nominatorsCount: 0,
-      depositsCount: 0,
-      withdrawalsCount: 0,
-      timestamp: new Date(),
-    });
-  }
-  return stats;
-}
-
-// Helper for StatsPerDomain
-export async function checkAndGetStatsPerDomain(
-  id: string,
-  blockNumber: number
-): Promise<StatsPerDomain> {
-  let statsPerDomain = await StatsPerDomain.get(id.toLowerCase());
-  if (!statsPerDomain) {
-    statsPerDomain = StatsPerDomain.create({
-      id: id.toLowerCase(),
-      domainId: "",
-      blockNumber: 0,
-      totalStaked: BigInt(0),
-      totalTaxCollected: BigInt(0),
-      totalRewardsCollected: BigInt(0),
-      totalDeposits: BigInt(0),
-      totalWithdrawals: BigInt(0),
-      totalShares: BigInt(0),
-      currentSharePrice: BigInt(0),
-      allTimeHighStaked: BigInt(0),
-      allTimeHighSharePrice: BigInt(0),
-      operatorsCount: 0,
-      activeOperatorsCount: 0,
-      slashedOperatorsCount: 0,
-      nominatorsCount: 0,
-      depositsCount: 0,
-      withdrawalsCount: 0,
-      timestamp: new Date(),
-    });
-  }
-  return statsPerDomain;
-}
-
-// Helper for StatsPerOperator
-export async function checkAndGetStatsPerOperator(
-  id: string,
-  blockNumber: number
-): Promise<StatsPerOperator> {
-  let statsPerOperator = await StatsPerOperator.get(id.toLowerCase());
-  if (!statsPerOperator) {
-    statsPerOperator = StatsPerOperator.create({
-      id: id.toLowerCase(),
-      domainId: "",
-      operatorId: "",
-      blockNumber: 0,
-      totalStaked: BigInt(0),
-      totalTaxCollected: BigInt(0),
-      totalRewardsCollected: BigInt(0),
-      totalDeposits: BigInt(0),
-      totalWithdrawals: BigInt(0),
-      totalShares: BigInt(0),
-      currentSharePrice: BigInt(0),
-      allTimeHighStaked: BigInt(0),
-      allTimeHighSharePrice: BigInt(0),
-      nominatorsCount: 0,
-      depositsCount: 0,
-      withdrawalsCount: 0,
-      timestamp: new Date(),
-    });
-  }
-  return statsPerOperator;
-}
-
-// Helper for StatsPerNominator
-export async function checkAndGetStatsPerNominator(
-  id: string,
-  blockNumber: number
-): Promise<StatsPerNominator> {
-  let statsPerNominator = await StatsPerNominator.get(id.toLowerCase());
-  if (!statsPerNominator) {
-    statsPerNominator = StatsPerNominator.create({
-      id: id.toLowerCase(),
-      domainId: "",
-      operatorId: "",
-      nominatorId: "",
-      blockNumber: 0,
-      totalStaked: BigInt(0),
-      totalDeposits: BigInt(0),
-      totalWithdrawals: BigInt(0),
-      totalShares: BigInt(0),
-      currentSharePrice: BigInt(0),
-      allTimeHighStaked: BigInt(0),
-      allTimeHighSharePrice: BigInt(0),
-      depositsCount: 0,
-      withdrawalsCount: 0,
-      timestamp: new Date(),
-    });
-  }
-  return statsPerNominator;
-}
-
-// Helper for StatsPerAccount
-export async function checkAndGetStatsPerAccount(
-  id: string,
-  blockNumber: number
-): Promise<StatsPerAccount> {
-  let statsPerAccount = await StatsPerAccount.get(id.toLowerCase());
-  if (!statsPerAccount) {
-    statsPerAccount = StatsPerAccount.create({
-      id: id.toLowerCase(),
-      accountId: "",
-      blockNumber: 0,
-      totalStaked: BigInt(0),
-      totalDeposits: BigInt(0),
-      totalWithdrawals: BigInt(0),
-      totalShares: BigInt(0),
-      currentSharePrice: BigInt(0),
-      allTimeHighStaked: BigInt(0),
-      allTimeHighSharePrice: BigInt(0),
-      operatorsCount: 0,
-      nominatorsCount: 0,
-      depositsCount: 0,
-      withdrawalsCount: 0,
-      timestamp: new Date(),
-    });
-  }
-  return statsPerAccount;
 }
